# tests_data/openapi_pre_destructure.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SentinelNet API",
    "version": "0.2.0-beta.1"
  },
  "paths": {
...
    "/api/local-devices": {
      "get": {
        "tags": [
          "Inventory"
        ],
        "summary": "Get Devices And Versions",
        "operationId": "get_devices_and_versions_api_local_devices_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/export/devices": {
      "get": {
        "tags": [
          "Inventory"
        ],
        "summary": "Export Devices Csv",
        "operationId": "export_devices_csv_api_export_devices_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/add-device": {
      "post": {
        "tags": [
          "Inventory"
        ],
        "summary": "Add Device",
        "operationId": "add_device_api_add_device_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/delete-device": {
      "post": {
        "tags": [
          "Inventory"
        ],
        "summary": "Delete Device",
        "operationId": "delete_device_api_delete_device_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceDelete"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/rename-device": {
      "post": {
        "tags": [
          "Inventory"
        ],
        "summary": "Rename Device",
        "description": "Rinomina un dispositivo gestito impostandone manualmente l'hostname (il\nnome mostrato in inventario e sulla mappa). admin/operator, con scoping.",
        "operationId": "rename_device_api_rename_device_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceRenameSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/import-csv": {
      "post": {
        "tags": [
          "Inventory"
        ],
        "summary": "Import Csv",
        "operationId": "import_csv_api_import_csv_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CSVImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/promote-device": {
      "post": {
        "tags": [
          "Inventory"
        ],
        "summary": "Promote Device",
        "description": "Promuove un dispositivo scoperto (CDP/LLDP) a dispositivo gestito,\naggiungendolo all'inventario così da poter essere sottoposto a triage.\nLe credenziali vanno completate dopo, nella pagina Inventario.",
        "operationId": "promote_device_api_promote_device_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromoteDeviceSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/reassign-device": {
      "post": {
        "tags": [
          "Inventory"
        ],
        "summary": "Reassign Device",
        "description": "Sposta un dispositivo in un gruppo diverso aggiornando solo il campo Group nel CSV.",
        "operationId": "reassign_device_api_reassign_device_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceReassignSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CSVImportRequest": {
        "properties": {
          "csv_data": {
            "type": "string",
            "title": "Csv Data"
          }
        },
        "type": "object",
        "required": [
          "csv_data"
        ],
        "title": "CSVImportRequest"
      },
...
      "DeviceDelete": {
        "properties": {
          "ip": {
            "type": "string",
            "title": "Ip"
          }
        },
        "type": "object",
        "required": [
          "ip"
        ],
        "title": "DeviceDelete"
      },
      "DeviceReassignSchema": {
        "properties": {
          "ip": {
            "type": "string",
            "title": "Ip"
          },
          "new_group": {
            "type": "string",
            "title": "New Group"
          }
        },
        "type": "object",
        "required": [
          "ip",
          "new_group"
        ],
        "title": "DeviceReassignSchema"
      },
      "DeviceRenameSchema": {
        "properties": {
          "ip": {
            "type": "string",
            "title": "Ip"
          },
          "hostname": {
            "type": "string",
            "title": "Hostname"
          }
        },
        "type": "object",
        "required": [
          "ip",
          "hostname"
        ],
        "title": "DeviceRenameSchema"
      },
      "DeviceSchema": {
        "properties": {
          "ip": {
            "type": "string",
            "pattern": "^\\d{1,3}\\.\\d{1,3}\\.\\d{1,3}\\.\\d{1,3}$",
            "title": "Ip"
          },
          "vendor": {
            "type": "string",
            "title": "Vendor"
          },
          "profile": {
            "type": "string",
            "title": "Profile"
          },
          "username": {
            "type": "string",
            "title": "Username",
            "default": "Admin"
          },
          "password": {
            "type": "string",
            "title": "Password",
            "default": "admin"
          },
          "enable_secret": {
            "type": "string",
            "title": "Enable Secret",
            "default": "admin"
          },
          "group": {
            "type": "string",
            "title": "Group",
            "default": "Generale"
          },
          "ssh_port": {
            "type": "integer",
            "maximum": 65535.0,
            "minimum": 1.0,
            "title": "Ssh Port",
            "default": 22
          },
          "transports": {
            "anyOf": [
              {
                "additionalProperties": {
                  "anyOf": [
                    {
                      "type": "integer"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transports"
          }
        },
        "type": "object",
        "required": [
          "ip",
          "vendor",
          "profile"
        ],
        "title": "DeviceSchema"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PromoteDeviceSchema": {
        "properties": {
          "node_id": {
            "type": "string",
            "title": "Node Id"
          },
          "ip": {
            "type": "string",
            "pattern": "^\\d{1,3}\\.\\d{1,3}\\.\\d{1,3}\\.\\d{1,3}$",
            "title": "Ip"
          },
          "vendor": {
            "type": "string",
            "title": "Vendor",
            "default": "cisco"
          },
          "group": {
            "type": "string",
            "title": "Group",
            "default": "Generale"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          },
          "device_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Type"
          },
          "hostname": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hostname"
          }
        },
        "type": "object",
        "required": [
          "node_id",
          "ip"
        ],
        "title": "PromoteDeviceSchema"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}